# Avoid unnecessary allocations on each cluster action