No change in performance in the tests, but less allocated bytes and allocations in total.

```
name         old time/op    new time/op    delta
ClusterDo-8    5.59µs ± 1%    5.60µs ± 3%     ~     (p=0.434 n=9+10)

name         old alloc/op   new alloc/op   delta
ClusterDo-8    4.85kB ± 0%    4.78kB ± 0%   -1.37%  (p=0.000 n=9+9)

name         old allocs/op  new allocs/op  delta
ClusterDo-8      20.0 ± 0%      17.0 ± 0%  -15.00%  (p=0.000 n=10+10)
```

Fixes #32